+++ hs/src/HelVM/HelPA/Assemblers/Frontend/HPL/Instruction.hs
@@ -1,0 +1,7 @@
+module HelVM.HelPA.Assemblers.Frontend.HPL.Instruction where
+
+data Element
+  = Asm { inline :: Bool, name :: Text, args :: [Text], body :: Text }
+  | Def { inline :: Bool, name :: Text, args :: [Text], body :: Text }
+  | Section { title :: Text }
+  deriving stock (Show)